WIP(zones): fmt & parse TTL field

+++ src/dns/zones/mod.rs
@@ -1,14 +1,13 @@
 pub(crate) mod zone_parser;
 
 use crate::dns::records::DNSRecord;
-use crate::dns::zones::zone_parser::zone_parser;
 use std::collections::HashMap;
 
 #[derive(Debug, PartialEq)]
 pub struct Zone {
     pub origin: Option<String>,
     pub name: String,
-    pub ttl: u32,
+    pub ttl: u16,
     pub soa: Option<DNSRecord>,
     pub records: HashMap<String, Vec<DNSRecord>>,
 }
